+++ vol1/csharp-examples/AIFH-Vol1-Core/Core/Randomize/AbstractGenerateRandom.cs
@@ -64,7 +64,7 @@ namespace AIFH_Vol1.Core.Randomize
         public abstract bool NextBoolean();
 
         /// <inheritdoc/>
-        public abstract double NextFloat();
+        public abstract float NextFloat();
 
         /// <inheritdoc/>
         public abstract double NextGaussian();
