+++ vol1/csharp-examples/AIFH-Vol1-Core/Core/Randomize/BasicGenerateRandom.cs
@@ -69,7 +69,7 @@ namespace AIFH_Vol1.Core.Randomize
         }
 
         /// <inheritdoc/>
-        public override double NextFloat()
+        public override float NextFloat()
         {
             return (float)_rnd.NextDouble();
         }
