+++ vol1/csharp-examples/AIFH-Vol1-Core/Core/Randomize/MultiplyWithCarryGenerateRandom.cs
@@ -158,7 +158,7 @@ namespace AIFH_Vol1.Core.Randomize
         /**
          * {@inheritDoc}
          */
-        public override double NextFloat()
+        public override float NextFloat()
         {
             return (float)NextDouble();
         }
